Student_project/requirements.txt
arabic-reshaper==3.0.0
asgiref==3.8.1
asn1crypto==1.5.1
certifi==2024.6.2
cffi==1.17.1
channels==4.1.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
crispy-bootstrap5==2024.2
cryptography==43.0.1
cssselect2==0.7.0
dj-stripe==2.8.4
Django==5.0.6
django-crispy-forms==2.1
django-model-utils==4.5.1
django-notifications-hq==1.8.3
django-rest==0.8.7
django-tinymce==4.0.0
djangorestframework==3.15.2
gunicorn==22.0.0
html5lib==1.1
idna==3.7
jsonfield==3.1.0
lxml==5.3.0
oscrypto==1.3.0
packaging==24.1
pillow==10.3.0
pycparser==2.22
pyHanko==0.25.1
pyhanko-certvalidator==0.26.3
pypdf==5.0.0
pypng==0.20220715.0
python-bidi==0.6.0
pytz==2024.1
PyYAML==6.0.2
qrcode==7.4.2
razorpay==1.4.2
reportlab==4.2.2
requests==2.32.3
setuptools==72.1.0
six==1.16.0
sqlparse==0.5.0
stripe==4.2.0
svglib==1.5.1
swapper==1.4.0
tinycss2==1.3.0
typing_extensions==4.12.2
tzdata==2024.1
tzlocal==5.2
uritools==4.0.3
urllib3==2.2.2
webencodings==0.5.1
xhtml2pdf==0.2.16
